More TK and dropping sanity fixes (#36475)
* forgot these

* cleaner

* better check

* fixes logic here, proc is only used now anyways

* null check

* fixes this not going up right

* redundant check

* cuts these down

* comma

* FINALLY fixes this proc

* can't drop checks

* updates these

* some more

* starts cutting down this line EVERYWHERE

* fixes these

* more cutdown

* last of em

* don't make this output "1"

* cleaner

* cuts down brick soap code

* sanity

* fixes this from creating it anyways, also cuts down use of new

* fixes this back

* fixes this a bit, still not working yet

* ornaments don't need to make a new one, also transfers dropping vars

* cuts down bluntcode while converting to create in hands

* cuts down circuitboard code

* fix grammar

* cuts down ALL of these!

* cuts these down too

* removes legacy before take item stuff by... cutting down code!

* makes this work again

* lines no longer needed, handled here now

* forgot this

* bot code cutdown

+++ code/modules/mob/inventory.dm
@@ -220,7 +220,8 @@
 	if(held_items[index])
 		return 0
 
-	if(!Adjacent(W) && isturf(W.loc) && !W.arcanetampered)
+	// adjacent only loc checks one way around, so first part is necessary, also for anything inside too
+	if(W.loc && !W.recursive_in_contents_of(src) && !Adjacent(W) && !W.arcanetampered)
 		return 0
 
 	if((W.flags & MUSTTWOHAND) && !(M_STRONG in mutations))
@@ -261,25 +262,57 @@
 		return 1
 	if(put_in_inactive_hand(W))
 		return 1
-	W.forceMove(isatom(W.loc) ? get_turf(W) : get_turf(src))
+	W.forceMove(W.loc ? get_turf(W) : get_turf(src))
 	W.reset_plane_and_layer()
 	W.dropped()
 	return 0
 
 //Helper proc tied to above for creating something in-hand via construction
-/mob/proc/create_in_hands(var/obj/item/olditem, var/obj/item/newitem, var/obj/item/using, var/uses = 1)
+/mob/proc/create_in_hands(var/obj/item/olditem, var/obj/item/newitem, var/obj/item/using, var/uses = 1, var/msg, var/vismsg, var/sound, var/move_in = FALSE)
 	if(!olditem || !newitem)
 		return 0
-	if(olditem.loc == src)
-		drop_item(olditem, force_drop = 1) // Necessary to show up in the same hand for below
-		. = put_in_hands(newitem)
-	qdel(olditem)
+	. = 0
+	if(ispath(newitem))
+		newitem = new newitem(olditem.loc)
+	if(!newitem) // If it deleted after creation
+		return 0
 	if(using)
 		if(istype(using,/obj/item/stack) && uses)
 			var/obj/item/stack/S = using
+			if(S.amount < uses)
+				to_chat(src, "<span class='warning'>You need at least [S.amount] [S.correct_name()] to make [newitem]!")
+				qdel(newitem)
+				return 0
 			S.use(uses)
+		else if(move_in)
+			if(using.loc != src)
+				using.forceMove(newitem)
+			else if(!drop_item(using, newitem, failmsg = TRUE))
+				qdel(newitem)
+				return 0
 		else
+			if(using.loc == src && !drop_item(using, failmsg = TRUE))
+				qdel(newitem)
+				return 0
 			qdel(using)
+		//using.transfer_fingerprints_to(newitem)
+	if(olditem.loc == src)
+		drop_item(olditem, force_drop = 1) // Necessary to show up in the same hand for below
+		. = put_in_hands(newitem)
+	newitem.cant_drop = olditem.cant_drop
+	newitem.canremove = olditem.canremove
+	if(move_in)
+		olditem.forceMove(newitem)
+	else
+		qdel(olditem)
+	if(vismsg)
+		visible_message(vismsg, msg)
+	else if(msg)
+		to_chat(src, msg)
+	if(sound)
+		playsound(src, sound, 50, 1)
+	if(olditem.loc != src)
+		. = 1 // return this if not runtiming
 
 /mob/proc/set_hand_amount(new_amount)
 	if(new_amount < held_items.len) //Decrease hand amount - drop items held in hands which will no longer exist!
@@ -331,9 +364,14 @@
 
 //Drops the item in our hand - you can specify an item and a location to drop to
 
-/mob/proc/drop_item(var/obj/item/to_drop, var/atom/Target, force_drop = 0) //Set force_drop to 1 to force the item to drop (even if it can't be dropped normally)
+/mob/proc/drop_item(var/obj/item/to_drop, var/atom/Target, force_drop = 0, failmsg) //Set force_drop to 1 to force the item to drop (even if it can't be dropped normally)
 
+	if(failmsg == TRUE)
+		failmsg = to_drop ? "<span class='warning'>You can't let go of \the [to_drop]!</span>" : null
+		
 	if(!candrop) //can't drop items while etheral
+		if(failmsg)
+			to_chat(src, failmsg)
 		return 0
 
 	if(!to_drop) //if we're not told to drop something specific
@@ -343,6 +381,8 @@
 		return 0 //bail
 
 	if((to_drop.cant_drop > 0) && !force_drop)
+		if(failmsg)
+			to_chat(src, failmsg)
 		return 0
 
 	if(!Target)
@@ -370,6 +410,8 @@
 				if(L)
 					L.check(list(to_drop))
 		return 1
+	if(failmsg)
+		to_chat(src, failmsg)
 	return 0
 
 /mob/proc/drop_hands(var/atom/Target, force_drop = 0) //drops both items
